// nautilus-ionic/src/app/api/model/inventoryItem.ts
/**
 * nautilus-central-service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InventoryItemCategory } from './inventoryItemCategory';


export interface InventoryItem { 
    id?: number;
    name?: string;
    itemCategory?: InventoryItemCategory;
}